requirements.txt
fastapi
uvicorn
numpy
soundfile
torch
torchaudio
openai-whisper
silero-vad
noisereduce
